🧬 Pull the functions about processing dividing serialization up to base class and adjust the usage of implementations.
